% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cytfly.R
\name{Cytfly}
\alias{Cytfly}
\title{Cytfly - 通讯概率差异分析 (审美定制版)}
\usage{
Cytfly(
  seurat_obj,
  target_gene,
  species,
  annotation_col = "seurat_clusters",
  output_dir = "./Cytfly_output"
)
}
\description{
Cytfly - 通讯概率差异分析 (审美定制版)
}
